#47 refactor(performance): 공연 상세 조회 API - DTO 패키지 구조 변경 및 DTO 일반 클래스 record로 변경
