+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/dto/performance/performancedetail/response/PerformanceScheduleResponse.java
@@ -1,0 +1,10 @@
+package wisoft.nextframe.schedulereservationticketing.dto.performance.performancedetail.response;
+
+import java.util.UUID;
+
+import lombok.Builder;
+
+@Builder
+public record PerformanceScheduleResponse(UUID id, String date, String time) {
+
+}
