+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/dto/performance/performancedetail/response/SeatSectionPriceResponse.java
@@ -1,0 +1,8 @@
+package wisoft.nextframe.schedulereservationticketing.dto.performance.performancedetail.response;
+
+import lombok.Builder;
+
+@Builder
+public record SeatSectionPriceResponse(String section, Integer price) {
+
+}
