+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/dto/performance/performancedetail/response/StadiumResponse.java
@@ -1,0 +1,9 @@
+package wisoft.nextframe.schedulereservationticketing.dto.performance.performancedetail.response;
+
+import java.util.UUID;
+
+import lombok.Builder;
+
+@Builder
+public record StadiumResponse(UUID id, String name, String address) {
+}
