+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/dto/performance/performancelist/response/PerformanceResponse.java
@@ -1,4 +1,4 @@
-package wisoft.nextframe.schedulereservationticketing.dto.performancelist.response;
+package wisoft.nextframe.schedulereservationticketing.dto.performance.performancelist.response;
 
 import java.time.LocalDate;
 import java.util.Date;
