--- schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/dto/performancedetail/response/SeatSectionPriceResponse.java
@@ -1,12 +1,0 @@
-package wisoft.nextframe.schedulereservationticketing.dto.performancedetail.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class SeatSectionPriceResponse {
-
-	private final String section;
-	private final Integer price;
-}
