+++ schedule-reservation-ticketing/src/main/java/wisoft/nextframe/schedulereservationticketing/service/performance/PerformanceService.java
@@ -13,13 +13,13 @@ import org.springframework.transaction.annotation.Transactional;
 
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
-import wisoft.nextframe.schedulereservationticketing.dto.performancedetail.response.PerformanceDetailResponse;
-import wisoft.nextframe.schedulereservationticketing.dto.performancedetail.response.PerformanceScheduleResponse;
-import wisoft.nextframe.schedulereservationticketing.dto.performancedetail.response.SeatSectionPriceResponse;
-import wisoft.nextframe.schedulereservationticketing.dto.performancedetail.response.StadiumResponse;
-import wisoft.nextframe.schedulereservationticketing.dto.performancelist.response.PaginationResponse;
-import wisoft.nextframe.schedulereservationticketing.dto.performancelist.response.PerformanceListResponse;
-import wisoft.nextframe.schedulereservationticketing.dto.performancelist.response.PerformanceResponse;
+import wisoft.nextframe.schedulereservationticketing.dto.performance.performancedetail.response.PerformanceDetailResponse;
+import wisoft.nextframe.schedulereservationticketing.dto.performance.performancedetail.response.PerformanceScheduleResponse;
+import wisoft.nextframe.schedulereservationticketing.dto.performance.performancedetail.response.SeatSectionPriceResponse;
+import wisoft.nextframe.schedulereservationticketing.dto.performance.performancedetail.response.StadiumResponse;
+import wisoft.nextframe.schedulereservationticketing.dto.performance.performancelist.response.PaginationResponse;
+import wisoft.nextframe.schedulereservationticketing.dto.performance.performancelist.response.PerformanceListResponse;
+import wisoft.nextframe.schedulereservationticketing.dto.performance.performancelist.response.PerformanceResponse;
 import wisoft.nextframe.schedulereservationticketing.entity.performance.Performance;
 import wisoft.nextframe.schedulereservationticketing.entity.schedule.Schedule;
 import wisoft.nextframe.schedulereservationticketing.entity.stadium.Stadium;
